+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## [Unreleased] - Next release 0.101.2 or 1.0.0
 
 ### Added
+- Memory profiler for Monitor when at DEBUG log level
 
 ### Changed
 
@@ -14,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Removed
 
 ### Fixed
+- Fixed Mongo Connection thread issue in Monitor
 
 ### Tech-Debt/Refactor